+++ include/ytp/control.h
@@ -15,53 +15,7 @@
 /**
  * @file control.h
  * @date 23 Apr 2021
- * @brief File contains C declaration of control layer of YTP.\n
- * A control channel is used for communicating peer, channel, publisher and
- * subscription control information.
- * A publisher is a peer that may publish messages on a channel.
- *
- * @par Description
- * - Control message\n
- * <table>
- * <caption id="multi_row">Control message</caption>
- * <tr><th colspan="3">peer/channel/time header  <th>control
- * <tr><th>8 bytes  <th>8 bytes  <th>8 bytes  <th>variable
- * <tr><td>Peer ID  <td>Channel ID  <td>Timestmap  <td>Payload
- * </table>
- * - Channel message\n
- * A peer announces a channel.\n
- * <table>
- * <caption id="multi_row">Channel message</caption>
- * <tr><th colspan="3">peer/channel/time header  <th>channel announcement
- * payload <tr><th>8 bytes  <th>8 bytes  <th>8 bytes  <th>variable <tr><td>Peer
- * ID  <td>Ctrl Channel ID = 0  <td>Timestmap  <td>Channel name
- * </table>
- * The peer needs to avoid publishing duplicated channel messages if the last
- * channel message in yamal has the same payload.\n
- * - Subscribe message\n
- * A peer announces a subscription to a channel.\n
- * <table>
- * <caption id="multi_row">Subscribe message</caption>
- * <tr><th colspan="3">peer/channel/time  <th>subscribe payload
- * <tr><th>8 bytes  <th>8 bytes  <th>8 bytes  <th>variable
- * <tr><td>Peer ID  <td>Ctrl Channel ID = 1  <td>Timestmap  <td>Payload
- * </table>
- * - Payload: If it doesn't end with character "/", a channel name; otherwise, a
- * prefix.\n
- * The peer needs to avoid publishing duplicated subscribe messages if the last
- * subscription message in yamal has the same payload.
- * - Directory message\n
- * Describes the messages that will be published on a particular channel.\n
- * <table>
- * <caption id="multi_row">Directory message</caption>
- * <tr><th colspan="3">peer/channel/time header  <th>directory payload
- * <tr><th>8 bytes  <th>8 bytes  <th>8 bytes  <th>variable
- * <tr><td>Peer ID  <td>Ctrl Channel ID = 2  <td>Timestmap  <td>Payload
- * </table>
- * - Payload : An Simple Channel Directory Protocol (SCDP) encoded string. This
- * field can be empty.\n
- * The peer needs to avoid publishing duplicated directory message if the last
- * directory message in yamal has the same payload.
+ * @brief File contains C declaration of control layer of YTP.
  *
  * @see http://www.featuremine.com
  */
@@ -118,7 +72,7 @@ FMMODFUNC void ytp_control_init(ytp_control_t *ctrl, fmc_fd fd,
  * @brief Allocates and initializes a ytp_control_t object
  *
  * @param[in] fd a yamal file descriptor
- * @param[in] enable_thread enable the preallocation and sync thread
+ * @param[in] enable_thread enables the auxiliary thread
  * @param[out] error out-parameter for error handling
  * @return ytp_control_t object
  */
@@ -130,7 +84,7 @@ FMMODFUNC ytp_control_t *ytp_control_new_2(fmc_fd fd, bool enable_thread,
  *
  * @param[in] ctrl the ytp_control_t object
  * @param[in] fd a yamal file descriptor
- * @param[in] enable_thread enable the preallocation and sync thread
+ * @param[in] enable_thread enables the auxiliary thread
  * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_control_init_2(ytp_control_t *ctrl, fmc_fd fd,
@@ -174,7 +128,7 @@ FMMODFUNC char *ytp_control_reserve(ytp_control_t *ctrl, size_t sz,
  * @param[in] msgtime the time to publish the data
  * @param[in] data the value returned by ytp_control_reserve()
  * @param[out] error out-parameter for error handling
- * @return iterator to the next memory mapped node
+ * @return ytp_iterator_t for the message
  */
 FMMODFUNC ytp_iterator_t ytp_control_commit(ytp_control_t *ctrl,
                                             ytp_peer_t peer,
@@ -185,11 +139,9 @@ FMMODFUNC ytp_iterator_t ytp_control_commit(ytp_control_t *ctrl,
 /**
  * @brief Publishes a subscription message
  *
- * Complexity: Constant on average, worst case linear in the size of the list.
- *
  * @param[in] ctrl the ytp_control_t object
  * @param[in] peer the peer that publishes the subscription message
- * @param[in] time the time to publish the subscription message
+ * @param[in] msgtime the time to publish the subscription message
  * @param[in] sz size of the payload
  * @param[in] payload a prefix or channel name
  * @param[out] error out-parameter for error handling
@@ -201,11 +153,9 @@ FMMODFUNC void ytp_control_sub(ytp_control_t *ctrl, ytp_peer_t peer,
 /**
  * @brief Publishes a directory message
  *
- * Complexity: Constant on average, worst case linear in the size of the list.
- *
  * @param[in] ctrl the ytp_control_t object
  * @param[in] peer the peer that publishes the directory message
- * @param[in] time the time to publish the directory message
+ * @param[in] msgtime the time to publish the directory message
  * @param[in] sz size of the payload
  * @param[in] payload a SCDP encoded string
  * @param[out] error out-parameter for error handling
@@ -240,7 +190,7 @@ FMMODFUNC void ytp_control_ch_name(ytp_control_t *ctrl, ytp_channel_t channel,
  * @param[in] sz size of the channel name
  * @param[in] name name of the channel
  * @param[out] error out-parameter for error handling
- * @return channel reference declared
+ * @return channel id
  */
 FMMODFUNC ytp_channel_t ytp_control_ch_decl(ytp_control_t *ctrl,
                                             ytp_peer_t peer, uint64_t msgtime,
@@ -271,15 +221,14 @@ FMMODFUNC void ytp_control_peer_name(ytp_control_t *ctrl, ytp_peer_t peer,
  * @param[in] sz size of the peer name
  * @param[in] name name of the peer
  * @param[out] error out-parameter for error handling
- * @return peer reference declared
+ * @return peer id
  */
 FMMODFUNC ytp_peer_t ytp_control_peer_decl(ytp_control_t *ctrl, size_t sz,
                                            const char *name,
                                            fmc_error_t **error);
 
 /**
- * @brief Finds next message on control level, moves iterator forward if there
- * is a next message.
+ * @brief moves iterator forward
  *
  * @param[in] ctrl the ytp_control_t object
  * @param[in] iter input iterator
@@ -292,9 +241,6 @@ FMMODFUNC ytp_iterator_t ytp_control_next(ytp_control_t *ctrl,
 
 /**
  * @brief Reads a message on control level
- *
- * Reads a message on a control level indicating in the case of peer
- * and channel announcement whether announcements are duplicated.
  *
  * @param[in] ctrl the ytp_control_t object
  * @param[in] iter iterator that points to the message node to read from
@@ -322,7 +268,6 @@ FMMODFUNC ytp_iterator_t ytp_control_begin(ytp_control_t *ctrl,
 
 /**
  * @brief Returns the iterator to the end of the list, the last node.
- * Also moves control pointer to the end.
  *
  * @param[in] ctrl the ytp_control_t object
  * @param[out] error out-parameter for error handling
@@ -340,14 +285,13 @@ FMMODFUNC bool ytp_control_term(ytp_iterator_t iterator);
 
 /**
  * @brief Returns an iterator given a serializable ptr.
- * Also moves control pointer to catch up with iterator.
  *
  * @param[in] ctrl ytp_control_t object
  * @param[in] off the serializable ptr offset
  * @param[out] error out-parameter for error handling
  * @return the iterator of the serializable ptr
  */
-FMMODFUNC ytp_iterator_t ytp_control_seek(ytp_control_t *ctrl, size_t off,
+FMMODFUNC ytp_iterator_t ytp_control_seek(ytp_control_t *ctrl, uint64_t off,
                                           fmc_error_t **error);
 
 /**
@@ -358,7 +302,7 @@ FMMODFUNC ytp_iterator_t ytp_control_seek(ytp_control_t *ctrl, size_t off,
  * @param[out] error out-parameter for error handling
  * @return the serializable ptr offset
  */
-FMMODFUNC size_t ytp_control_tell(ytp_control_t *ctrl, ytp_iterator_t iterator,
+FMMODFUNC uint64_t ytp_control_tell(ytp_control_t *ctrl, ytp_iterator_t iterator,
                                   fmc_error_t **error);
 
 #ifdef __cplusplus
