+++ include/ytp/yamal.h
@@ -49,7 +49,7 @@ typedef enum { CLOSABLE = 1, UNCLOSABLE = 2 } FMC_CLOSABLE;
  *
  * @param[out] yamal
  * @param[in] fd a yamal file descriptor
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_yamal_init(ytp_yamal_t *yamal, int fd, fmc_error_t **error);
 
@@ -57,7 +57,7 @@ FMMODFUNC void ytp_yamal_init(ytp_yamal_t *yamal, int fd, fmc_error_t **error);
  * @brief Allocates and initializes a ytp_yamal_t object
  *
  * @param[in] fd a yamal file descriptor
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_yamal_t object
  */
 FMMODFUNC ytp_yamal_t *ytp_yamal_new(int fd, fmc_error_t **error);
@@ -79,8 +79,8 @@ FMMODFUNC void ytp_yamal_clear_aux_thread_affinity();
  *
  * @param[out] yamal
  * @param[in] fd a yamal file descriptor
- * @param[in] enable_thread enable the preallocation and sync thread
- * @param[out] error
+ * @param[in] enable_thread enables the auxiliary thread
+ * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_yamal_init_2(ytp_yamal_t *yamal, int fd, bool enable_thread,
                                 fmc_error_t **error);
@@ -89,8 +89,8 @@ FMMODFUNC void ytp_yamal_init_2(ytp_yamal_t *yamal, int fd, bool enable_thread,
  * @brief Allocates and initializes a ytp_yamal_t object
  *
  * @param[in] fd a yamal file descriptor
- * @param[in] enable_thread enable the preallocation and sync thread
- * @param[out] error
+ * @param[in] enable_thread enables the auxiliary thread
+ * @param[out] error out-parameter for error handling
  * @return ytp_yamal_t object
  */
 FMMODFUNC ytp_yamal_t *ytp_yamal_new_2(int fd, bool enable_thread,
@@ -100,9 +100,9 @@ FMMODFUNC ytp_yamal_t *ytp_yamal_new_2(int fd, bool enable_thread,
  * @brief Initializes a ytp_yamal_t object
  *
  * @param[in] fd a yamal file descriptor
- * @param[in] enable_thread enable the preallocation and sync thread
+ * @param[in] enable_thread enables the auxiliary thread
  * @param[in] closable closable mode
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_yamal_t object
  */
 FMMODFUNC void ytp_yamal_init_3(ytp_yamal_t *yamal, int fd, bool enable_thread,
@@ -112,9 +112,9 @@ FMMODFUNC void ytp_yamal_init_3(ytp_yamal_t *yamal, int fd, bool enable_thread,
  * @brief Allocates and initializes a ytp_yamal_t object
  *
  * @param[in] fd a yamal file descriptor
- * @param[in] enable_thread enable the preallocation and sync thread
+ * @param[in] enable_thread enables the auxiliary thread
  * @param[in] closable closable mode
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_yamal_t object
  */
 FMMODFUNC ytp_yamal_t *ytp_yamal_new_3(int fd, bool enable_thread,
@@ -124,7 +124,7 @@ FMMODFUNC ytp_yamal_t *ytp_yamal_new_3(int fd, bool enable_thread,
 /**
  * @brief Returns the file descriptor from a ytp_yamal_t object
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @return fmc_fd
  */
 FMMODFUNC fmc_fd ytp_yamal_fd(ytp_yamal_t *yamal);
@@ -132,9 +132,9 @@ FMMODFUNC fmc_fd ytp_yamal_fd(ytp_yamal_t *yamal);
 /**
  * @brief Reserves memory for data in the memory mapped list
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] sz the size of the data payload
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return a writable pointer for data
  */
 FMMODFUNC char *ytp_yamal_reserve(ytp_yamal_t *yamal, size_t sz,
@@ -143,10 +143,10 @@ FMMODFUNC char *ytp_yamal_reserve(ytp_yamal_t *yamal, size_t sz,
 /**
  * @brief Commits the data to the memory mapped list
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] data the value returned by ytp_yamal_reserve
  * @param[in] lstidx the list index to commit to
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_iterator_t for the message
  */
 FMMODFUNC ytp_iterator_t ytp_yamal_commit(ytp_yamal_t *yamal, void *data,
@@ -155,39 +155,39 @@ FMMODFUNC ytp_iterator_t ytp_yamal_commit(ytp_yamal_t *yamal, void *data,
 /**
  * @brief Reads a message on yamal level
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] iterator
  * @param[out] seqno
  * @param[out] sz
  * @param[out] data
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_yamal_read(ytp_yamal_t *yamal, ytp_iterator_t iterator,
-                              size_t *seqno, size_t *sz, const char **data,
+                              uint64_t *seqno, size_t *sz, const char **data,
                               fmc_error_t **error);
 
 /**
  * @brief Destroys a ytp_yamal_t object
  *
- * @param[in] yamal
- * @param[out] error
+ * @param[in] yamal the ytp_yamal_t object
+ * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_yamal_destroy(ytp_yamal_t *yamal, fmc_error_t **error);
 
 /**
  * @brief Destroys and deallocate a ytp_yamal_t object
  *
- * @param[in] yamal
- * @param[out] error
+ * @param[in] yamal the ytp_yamal_t object
+ * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_yamal_del(ytp_yamal_t *yamal, fmc_error_t **error);
 
 /**
  * @brief Returns an iterator to the beginning of the list
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] lstidx
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_iterator_t
  */
 FMMODFUNC ytp_iterator_t ytp_yamal_begin(ytp_yamal_t *yamal, size_t lstidx,
@@ -196,9 +196,9 @@ FMMODFUNC ytp_iterator_t ytp_yamal_begin(ytp_yamal_t *yamal, size_t lstidx,
 /**
  * @brief Returns an iterator to the end of the list
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] lstidx
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_iterator_t
  */
 FMMODFUNC ytp_iterator_t ytp_yamal_end(ytp_yamal_t *yamal, size_t lstidx,
@@ -215,9 +215,9 @@ FMMODFUNC bool ytp_yamal_term(ytp_iterator_t iterator);
 /**
  * @brief Returns the next iterator
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] iterator
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_iterator_t
  */
 FMMODFUNC ytp_iterator_t ytp_yamal_next(ytp_yamal_t *yamal,
@@ -227,9 +227,9 @@ FMMODFUNC ytp_iterator_t ytp_yamal_next(ytp_yamal_t *yamal,
 /**
  * @brief Returns the previous iterator
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] iterator
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_iterator_t
  */
 FMMODFUNC ytp_iterator_t ytp_yamal_prev(ytp_yamal_t *yamal,
@@ -239,38 +239,38 @@ FMMODFUNC ytp_iterator_t ytp_yamal_prev(ytp_yamal_t *yamal,
 /**
  * @brief Returns an iterator given a serializable ptr
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] ptr
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return ytp_iterator_t
  */
-FMMODFUNC ytp_iterator_t ytp_yamal_seek(ytp_yamal_t *yamal, size_t ptr,
+FMMODFUNC ytp_iterator_t ytp_yamal_seek(ytp_yamal_t *yamal, uint64_t ptr,
                                         fmc_error_t **error);
 
 /**
  * @brief Returns serializable ptr given an iterator
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] iterator
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  * @return serializable
  */
-FMMODFUNC size_t ytp_yamal_tell(ytp_yamal_t *yamal, ytp_iterator_t iterator,
+FMMODFUNC uint64_t ytp_yamal_tell(ytp_yamal_t *yamal, ytp_iterator_t iterator,
                                 fmc_error_t **error);
 
 /**
  * @brief Closes the yamal lists
  *
- * @param[in] yamal
- * @param[out] error
+ * @param[in] yamal the ytp_yamal_t object
+ * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_yamal_close(ytp_yamal_t *yamal, fmc_error_t **error);
 
 /**
  * @brief Determines if a list is closed
  *
- * @param[in] yamal
- * @param[out] error
+ * @param[in] yamal the ytp_yamal_t object
+ * @param[out] error out-parameter for error handling
  * @return true if the list is closed, false otherwise
  */
 FMMODFUNC bool ytp_yamal_closed(ytp_yamal_t *yamal, size_t lstidx,
@@ -279,9 +279,9 @@ FMMODFUNC bool ytp_yamal_closed(ytp_yamal_t *yamal, size_t lstidx,
 /**
  * @brief Allocates a specific page
  *
- * @param[in] yamal
+ * @param[in] yamal the ytp_yamal_t object
  * @param[in] page
- * @param[out] error
+ * @param[out] error out-parameter for error handling
  */
 FMMODFUNC void ytp_yamal_allocate_page(ytp_yamal_t *yamal, size_t page,
                                        fmc_error_t **error);
@@ -289,8 +289,8 @@ FMMODFUNC void ytp_yamal_allocate_page(ytp_yamal_t *yamal, size_t page,
 /**
  * @brief Returns the reserved size
  *
- * @param[in] yamal
- * @param[out] error
+ * @param[in] yamal the ytp_yamal_t object
+ * @param[out] error out-parameter for error handling
  * @return yamal size
  */
 FMMODFUNC size_t ytp_yamal_reserved_size(ytp_yamal_t *yamal,
